+++ BackEnd/Core/Boundaries/Persistence/IPublicationsRepository.cs
@@ -4,7 +4,7 @@ namespace Posterr.Core.Boundaries.Persistence;
 
 public interface IPublicationsRepository
 {
-    Task<ushort> CountPublicationsByUserBetween(IUser author, DateTime startInclusive, DateTime endInclusive);
+    Task<int> CountPublicationsByUserBetween(IUser author, DateTime startInclusive, DateTime endInclusive);
     Task<IPost?> FindPostById(long originalPostId);
     Task<IList<IPublication>> Paginate(int lastSeenRow, ushort pageSize);
     Task<IPost> PublishNewPost(IUnpublishedPost unpublishedPost);
